+++ glotaran/project/scheme.py
@@ -119,12 +119,12 @@ class Scheme:
         """
         return self.model.problem_list(self.parameters)
 
-    def validate(self) -> str:
+    def validate(self) -> MarkdownStr:
         """Return a string listing all problems in the model and missing parameters.
 
         Returns
         -------
-        str
+        MarkdownStr
             A user-friendly string containing all the problems of a model if any.
             Defaults to 'Your model is valid.' if no problems are found.
         """
@@ -150,9 +150,7 @@ class Scheme:
         """
         model_markdown_str = self.model.markdown(parameters=self.parameters)
 
-        markdown_str = "\n\n"
-        markdown_str += "__Scheme__\n\n"
-
+        markdown_str = "\n\n__Scheme__\n\n"
         if self.non_negative_least_squares is not None:
             markdown_str += f"* *non_negative_least_squares*: {self.non_negative_least_squares}\n"
         markdown_str += (