+++ changelog.md
@@ -27,6 +27,7 @@
 
 - 🩹 Fix Crash in optimization_group_calculator_linked when using guidance spectra (#950)
 - 🩹 ParameterGroup.get degrades full_label of nested Parameters with nesting over 2 (#1043)
+- 🩹 Show validation problem if parameters are missing values (default: NaN) (#1076)
 
 ### 📚 Documentation